requirement.txt
absl-py==0.9.0
astor==0.8.1
certifi==2019.11.28
cycler==0.10.0
decorator==4.4.1
gast==0.3.3
grpcio==1.26.0
h5py==2.10.0
imageio==2.6.1
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kungfu==0.2.1
Markdown==3.1.1
matplotlib==3.1.2
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
networkx==2.4
numpy==1.18.1
opencv-python==4.1.2.30
pandas==0.25.3
Pillow==7.0.0
protobuf==3.11.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
scikit-learn==0.22.1
scipy==1.3.2
six==1.14.0
sklearn==0.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.2
termcolor==1.1.0
tornado==6.0.3
tqdm==4.42.0
Werkzeug==0.16.1
